csProject/requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.13.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
billiard==3.3.0.23
binaryornot==0.4.4
blinker==1.4
celery==3.1.25
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.3.1
Deprecated==1.2.4
dominate==2.2.1
EditorConfig==0.12.2
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Bootstrap==3.3.6.0
Flask-Celery==2.4.3
Flask-Dance==1.1.0
Flask-Login==0.4.1
Flask-RQ==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
funcsigs==1.0.2
gitdb2==2.0.5
github3.py==1.2.0
GitPython==2.1.11
greenlet==0.4.15
gunicorn==19.9.0
humanize==0.5.1
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
jsbeautifier==1.7.5
jwcrypto==0.6.0
kombu==3.0.37
lazy==1.3
lxml==4.3.0
MarkupSafe==1.0
more-itertools==5.0.0
ndg-httpsclient==0.5.1
oauthlib==2.1.0
openpyxl==2.5.12
pathlib2==2.3.3
pluggy==0.8.1
py==1.7.0
pyasn1==0.4.4
pycount==0.6.18
pycparser==2.19
pygit2==0.27.2
PyGithub==1.43.3
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytest==4.1.1
pytest-flask==0.14.0
python-dateutil==2.7.5
pytz==2018.9
rabbitmq==0.2.0
redis==2.10.5
requests==2.20.1
requests-oauthlib==1.0.0
rq==0.6.0
scandir==1.9.0
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.2.13
SQLAlchemy-Utils==0.33.6
uritemplate==3.0.0
urllib3==1.24.1
URLObject==2.4.3
vine==1.2.0
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
